Add backtrack fixed point enumeration algorithm (#36)
Adds an alternative enumeration algorithm. This one tends to be faster (due to less bookkeeping overhead) when the lattice elements naturally decompose into several dimensions and this decomposition is consistent with the corresponding join and meet operations (e.g., because the join or meet operation essentially acts on each dimension independently). Specifically, the meet operation should have this structure if the fixed points of an increasing monotone function are sought, and the join operation should be like that for fixed points of a decreasing monotone function.

This alternative algorithm is made the standard choice for enumerating stable role equivalences w.r.t. role extensions on the lattice of equivalences and for stable role structures on the lattice of binary relations, as these are cases where this algorithm reveals its strengths.